RAG/requirements.txt
PyMuPDF
chromadb
rank-bm25
sentence-transformers
google-genai
python-dotenv
pytesseract
opencv-python
Pillow
numpy
